ios color picker

+++ ios/LEDLamp/LampState.swift
@@ -24,19 +24,19 @@ struct LampState {
         case invalidTemperature
     }
     
-    var red: UInt8
-    var green: UInt8
-    var blue: UInt8
-    var brightness: UInt8
+    var red: Double
+    var green: Double
+    var blue: Double
+    var brightness: Double
     var mode: Mode
     var temperature: ColorTemperature
     
     var data: Data {
         var buffer = Data()
-        buffer.append(red)
-        buffer.append(green)
-        buffer.append(blue)
-        buffer.append(brightness)
+        buffer.append(UInt8(red))
+        buffer.append(UInt8(green))
+        buffer.append(UInt8(blue))
+        buffer.append(UInt8(brightness))
         buffer.append(mode.rawValue)
         buffer.append(UInt8((temperature.rawValue >> 16) & 0xFF))
         buffer.append(UInt8((temperature.rawValue >> 8) & 0xFF))
@@ -58,6 +58,6 @@ struct LampState {
         offset += 1
         let temperature = Int(data[offset]) << 16 | Int(data[offset + 1]) << 8 | Int(data[offset + 2])
         guard let colorTemperature = ColorTemperature(rawValue: temperature) else { throw LampState.Error.invalidTemperature }
-        return LampState(red: red, green: green, blue: blue, brightness: brightness, mode: mode, temperature: colorTemperature)
+        return LampState(red: Double(red), green: Double(green), blue: Double(blue), brightness: Double(brightness), mode: mode, temperature: colorTemperature)
     }
 }
